requirements.txt
langchain
langchain-community
langchain-pinecone
langchain-openai
datasets
langchain-cohere
cohere
python-dotenv
flask